requirements.txt
pySerial == 3.5
keyboard == 0.13.5
tensorflow==2.13.0
numpy==1.24.3
pandas==2.0.3
wfdb==4.1.2
seaborn==0.12.2
matplotlib==3.7.3
keras==2.13.1
scikit-learn==1.3.1
